// src/types.ts
export interface ActivitySummary {
  id: number;
  name: string;
  distance: number;
  moving_time: number;
  elapsed_time: number;
  total_elevation_gain: number;
  type: string;
  sport_type: string;
  start_date: string;
  start_date_local: string;
  timezone: string;
  achievement_count: number;
  kudos_count: number;
  comment_count: number;
  athlete_count: number;
  average_speed: number;
  max_speed: number;
  average_heartrate?: number;
  max_heartrate?: number;
  average_watts?: number;
  max_watts?: number;
  weighted_average_watts?: number;
  kilojoules?: number;
  suffer_score?: number;
  average_cadence?: number;
  has_heartrate: boolean;
  gear_id: string | null;
  device_name?: string;
  calories?: number;
  description?: string;
  photos?: unknown;
  map?: {
    id: string;
    summary_polyline: string | null;
    polyline?: string | null;
  };
}
// ...
export interface GearSummary {
  id: string;
  primary: boolean;
  name: string;
  distance: number;
  brand_name?: string;
  model_name?: string;
  description?: string;
  frame_type?: number;
  resource_state?: number;
}
// ...
export interface Route {
  id: number;
  name: string;
  description: string;
  distance: number;
  elevation_gain: number;
  type: number;
  sub_type: number;
  starred: boolean;
  timestamp: number;
  map: {
    id: string;
    summary_polyline: string;
    polyline?: string;
  };
  estimated_moving_time: number;
}

/** Club */
export interface Club {
  id: number;
  name: string;
  profile_medium: string;
  profile: string;
  cover_photo: string;
  cover_photo_small: string;
  sport_type: string;
  activity_types: string[];
  city: string;
  state: string;
  country: string;
  private: boolean;
  member_count: number;
  featured: boolean;
  verified: boolean;
  url: string;
  description?: string;
}